requirements.txt
fastapi
httpx
pytest
python-dotenv
uvicorn
pytest-cov
